requirements/production.txt
django==5.1.3
gunicorn==23.0.0
whitenoise==6.8.2
psycopg==3.2.3
django-health-check==3.18.1
redis==5.2.0
setuptools==74.1.2
requests==2.32.3
django-ninja==1.3.0
openai==1.57.0
python-dotenv==1.0.1